arm64: imx8mp: sm2s: add checkboard into board_late_init
Add logic to read the following details from SM2S id-eeprom.
This is embedded into the new release of the SM2S smarc,
into the i2c eeprom we can see the following details:

 - feature .......... 24N0600I
 - serial ........... 1012213839
 - revision (MES) ... B0
 - ram size.......... 4GB
 - eMMC size......... 16GB

This would be helpfull in the future to load the right .dts
based on this details.

Signed-off-by: Tommaso Merciai <tomm.merciai@gmail.com>